+++ tests/dummy/app/locales/en/translations.js
@@ -684,7 +684,11 @@ $.extend(true, translations, {
             'settings-example': {
               'caption': 'Settings example',
               'title': ''
-            }
+            },
+            'flexberry-file-in-modal': {
+              'caption': 'Flexberry file in modal window',
+              'title': ''
+            },
           },
           'flexberry-groupedit': {
             'caption': 'flexberry-groupedit',