+++ tests/dummy/app/locales/ru/translations.js
@@ -704,7 +704,11 @@ $.extend(true, translations, {
             'settings-example': {
               'caption': 'Пример работы с настройками',
               'title': ''
-            }
+            },
+            'flexberry-file-in-modal': {
+              'caption': 'Пример файла в модальном окне',
+              'title': ''
+            },
           },
           'flexberry-groupedit': {
             'caption': 'flexberry-groupedit',